+++ mz_libcaer_OF/src/ABMOF/abmof_hw_accel.h
@@ -19,6 +19,6 @@
 #define BLOCK_SIZE 3
 #define SEARCH_DISTANCE 4
 
-void parseEvents(const uint32_t * data, int32_t eventsArraySize, int32_t *eventSlice);
+void parseEvents(const uint64_t * data, int32_t eventsArraySize, int32_t *eventSlice);
 
 #endif
